requirements.txt
# Bybit API
pybit>=5.7.0

# Конфигурация и окружение
python-dotenv>=1.0.0

# Удобная работа с данными и временем
pandas>=2.0.0
numpy>=1.24.0

# Управление ботом через Telegram
python-telegram-bot>=21.0
pytz  # для APScheduler/JobQueue (часы и отложенные задачи)

# Логирование (при желании можно заменить на loguru)
# loguru>=0.7.0
